+++ src/structures/vroom/job.cpp
@@ -23,7 +23,8 @@ Job::Job(Id id,
          const std::vector<TimeWindow>& tws,
          std::string description,
          const TypeToUserDurationMap& setup_per_type,
-         const TypeToUserDurationMap& service_per_type)
+         const TypeToUserDurationMap& service_per_type,
+         SoftTimeWindow soft_time_window)
   : location(location),
     id(id),
     type(JOB_TYPE::SINGLE),
@@ -36,7 +37,8 @@ Job::Job(Id id,
     tws(tws),
     description(std::move(description)),
     setup_per_type(utils::scale_from_user_duration(setup_per_type)),
-    service_per_type(utils::scale_from_user_duration(service_per_type)) {
+    service_per_type(utils::scale_from_user_duration(service_per_type)),
+    soft_time_window(std::move(soft_time_window)) {
   utils::check_tws(tws, id, "job");
   utils::check_priority(priority, id, "job");
   utils::check_no_empty_keys(this->setup_per_type, id, "job", "setup_per_type");
@@ -58,7 +60,8 @@ Job::Job(Id id,
          std::string description,
          const TypeToUserDurationMap& setup_per_type,
          const TypeToUserDurationMap& service_per_type,
-         std::string co_located_group)
+         std::string co_located_group,
+         SoftTimeWindow soft_time_window)
   : location(location),
     id(id),
     type(type),
@@ -72,7 +75,8 @@ Job::Job(Id id,
     description(std::move(description)),
     setup_per_type(utils::scale_from_user_duration(setup_per_type)),
     service_per_type(utils::scale_from_user_duration(service_per_type)),
-    co_located_group(std::move(co_located_group)) {
+    co_located_group(std::move(co_located_group)),
+    soft_time_window(std::move(soft_time_window)) {
   assert(type == JOB_TYPE::PICKUP || type == JOB_TYPE::DELIVERY);
   std::string type_str = (type == JOB_TYPE::PICKUP) ? "pickup" : "delivery";
   utils::check_tws(tws, id, type_str);